requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
opencv-python==4.9.0.80
numpy==1.26.3
python-multipart==0.0.6
PyJWT==2.8.0
